+++ app/src/main/java/dev/kichan/marketplace/model/service/CouponOwnerService.kt
@@ -2,7 +2,6 @@ package dev.kichan.marketplace.model.service
 
 import dev.kichan.marketplace.model.data.coupon.CouponPagenation
 import dev.kichan.marketplace.model.data.ResponseTemplate
-import dev.kichan.marketplace.model.data.coupon.ClosingCouponRes
 import dev.kichan.marketplace.model.data.coupon.CouponRes
 import dev.kichan.marketplace.model.data.coupon.CouponCreateReq
 import dev.kichan.marketplace.model.data.coupon.CouponUpdateReq
@@ -47,34 +46,4 @@ interface CouponOwnerService {
         @Body body: CouponCreateReq,
         @Query("marketId") marketId: Int,
     ): Response<ResponseTemplate<CouponRes>>
-}
-
-interface CouponService {
-    @GET("/api/coupons")
-    suspend fun getCouponList(
-        @Query("memberId") memberId: Long,
-        @Query("marketId") marketId: Long,
-        @Query("couponId") lastCouponId: Long?,
-        @Query("size") pageSize: Long?
-    ): Response<ResponseTemplate<CouponPagenation<CouponRes>>>
-
-    @GET("/api/coupons/popular")
-    suspend fun getPopularCoupon(
-        @Query("memberId") memberId: Long,
-        @Query("lastCouponId") lastCouponId: Long?,
-        @Query("pageSize") pageSize: Long?
-    ): Response<ResponseTemplate<CouponPagenation<CouponRes>>>
-
-    @GET("/api/coupons/latest")
-    suspend fun getLatestCoupon(
-        @Query("memberId") memberId: Long,
-        @Query("lastCreatedAt") lastCreatedAt: String?,
-        @Query("lastCouponId") lastCouponId: Long?,
-        @Query("pageSize") pageSize: Int
-    ): Response<ResponseTemplate<CouponPagenation<CouponRes>>>
-
-    @GET("/api/coupons/closing")
-    suspend fun getClosingCoupon(
-        @Query("pageSize") pageSize: Int
-    ): Response<ResponseTemplate<List<ClosingCouponRes>>>
 }
